+++ common/erebus/item/ItemPortalActivator.java
@@ -52,7 +52,7 @@ public class ItemPortalActivator extends Item {
 			int var11 = par3World.getBlockId(par4, par5, par6);
 
 			if (var11 == 0) {
-				par3World.playSoundEffect((double) par4 + 0.5D, (double) par5 + 0.5D, (double) par6 + 0.5D, "fire.ignite", 1.0F, itemRand.nextFloat() * 0.4F + 0.8F);
+				par3World.playSoundEffect(par4 + 0.5D, par5 + 0.5D, par6 + 0.5D, "fire.ignite", 1.0F, itemRand.nextFloat() * 0.4F + 0.8F);
 				onBlockAdded(par3World, par4, par5, par6);
 			}
