+++ common/erebus/world/experimental/GenLayerErebusSwampRivers.java
@@ -21,7 +21,7 @@ public class GenLayerErebusSwampRivers extends GenLayerErebus {
 
 		for (int var7 = 0; var7 < par4; ++var7) {
 			for (int var8 = 0; var8 < par3; ++var8) {
-				this.initChunkSeed((long) (var8 + par1), (long) (var7 + par2));
+				this.initChunkSeed(var8 + par1, var7 + par2);
 				int var9 = var5[var8 + 1 + (var7 + 1) * (par3 + 2)];
 
 				if ((var9 != BiomeGenBase.swampland.biomeID || this.nextInt(6) != 0) && (var9 != BiomeGenBase.jungle.biomeID && var9 != BiomeGenBase.jungleHills.biomeID || this.nextInt(8) != 0)) {
